Refactor les fonctions d'ellipses pour utiliser des types double et améliorer la précision

+++ libs/Ellipses.h
@@ -1,14 +1,16 @@
 #include "Common.h"
 
-float b(float a, float e);
-float c(float a, float e);
-float h(float a, float e);
-float f(float a, float e);
-float p(float a, float e);
+double b(double a, double e);
+double c(double a, double e);
+double h(double a, double e);
+double f(double a, double e);
+double p(double a, double e);
 
-float r(float TrueAnomaly, float p, float e);
+/*
+double r(double TrueAnomaly, double p, double e);
 
 float InitialX_2D(float TrueAnomaly, float Dist);
 float InitialY_2D(float TrueAnomaly, float Dist);
 float ActualX_2D(float x, float y, float PeAng);
 float ActualY_2D(float x, float y, float PeAng);
+*/
